control-plane/api/requirements.txt
alembic
fastapi[standard]
psycopg2-binary
sqlalchemy[asyncio]
asyncpg
aiosqlite
redis
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic
pydantic-settings
httpx
boto3
aiobotocore
aiofiles
pytest
pytest-asyncio
